changed DTOs from BigDecimal to Double

+++ src/main/java/io/contek/invoker/okx/api/common/_AccountBalance.java
@@ -1,7 +1,6 @@
 package io.contek.invoker.okx.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 import java.util.List;
 
 @NotThreadSafe
@@ -14,7 +13,7 @@ public class _AccountBalance {
   public String mmr; // Maintenance margin requirement in USD
   public String notionalUsd; // Notional value of positions in USD
   public String ordFroz; // Cross margin frozen for pending orders in USD
-  public BigDecimal totalEq; // The total amount of equity in USD
+  public Double totalEq; // The total amount of equity in USD
   public long uTime; // Update time of account information, millisecond format of Unix timestamp, e.g. 1597026383085
   public List<_EquityBalance> details;
 }
