+++ src/main/java/io/contek/invoker/okx/api/common/_AssetBalance.java
@@ -1,13 +1,12 @@
 package io.contek.invoker.okx.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _AssetBalance {
 
-  public BigDecimal availBal;
-  public BigDecimal bal;
+  public Double availBal;
+  public Double bal;
   public String ccy;
-  public BigDecimal frozenBal;
+  public Double frozenBal;
 }
