+++ src/main/java/io/contek/invoker/okx/api/common/_FeeRate.java
@@ -1,14 +1,13 @@
 package io.contek.invoker.okx.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _FeeRate {
 
     public String level;
-    public BigDecimal taker;
-    public BigDecimal maker;
+    public Double taker;
+    public Double maker;
     public String takerU;
     public String makerU;
     public String delivery;
