+++ src/main/java/io/contek/invoker/okx/api/common/_InterestRate.java
@@ -1,11 +1,10 @@
 package io.contek.invoker.okx.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _InterestRate {
 
     public String ccy;
-    public BigDecimal interestRate;
+    public Double interestRate;
 }
