+++ src/main/java/io/contek/invoker/okx/api/common/_Leverage.java
@@ -1,7 +1,6 @@
 package io.contek.invoker.okx.api.common;
 
 import javax.annotation.concurrent.NotThreadSafe;
-import java.math.BigDecimal;
 
 @NotThreadSafe
 public class _Leverage {
@@ -9,5 +8,5 @@ public class _Leverage {
   public String instId;
   public String mgnMode;
   public String posSide;
-  public BigDecimal lever;
+  public Double lever;
 }
